Refactor: extract geometry helpers
